trunk->0.3 backport: r6563 [2018522] Fixed Lua marker deserialisation bug caused by unmarshallLong not expecting 64-bit longs.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/branches/stone_soup-0.4@6571 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/tags.h
@@ -13,7 +13,8 @@
 #ifndef TAGS_H
 #define TAGS_H
 
-#include <stdio.h>
+#include <cstdio>
+#include <stdint.h>
 #include "externs.h"
 
 enum tag_type   // used during save/load process to identify data blocks
@@ -85,8 +86,8 @@ private:
 };
 
 void marshallByte    (writer &, char );
-void marshallShort   (writer &, short );
-void marshallLong    (writer &, long );
+void marshallShort   (writer &, int16_t );
+void marshallLong    (writer &, int32_t );
 void marshallFloat   (writer &, float );
 void marshallBoolean (writer &, bool );
 void marshallString  (writer &, const std::string &, int maxSize = 0);
@@ -116,8 +117,8 @@ private:
 };
 
 char        unmarshallByte    (reader &);
-short       unmarshallShort   (reader &);
-long        unmarshallLong    (reader &);
+int16_t     unmarshallShort   (reader &);
+int32_t     unmarshallLong    (reader &);
 float       unmarshallFloat   (reader &);
 bool        unmarshallBoolean (reader &);
 int         unmarshallCString (reader &, char *data, int maxSize);
